some fixes for running with java 5

+++ cache/trunk/coconut-cache-impl/src/test/java/org/coconut/cache/internal/service/statistics/AtomicDoubleTest.java
@@ -219,7 +219,7 @@ public class AtomicDoubleTest extends TestCase {
         AtomicDouble ai = new AtomicDouble();
         for (int i = -12; i < 6; ++i) {
             ai.set(i);
-            assertEquals((long) i, ai.longValue());
+            assertEquals(i, ai.longValue());
         }
     }
