+++ cache/trunk/coconut-cache-test/coconut-cache-test-framework/src/main/java/org/coconut/cache/test/keys/KeyValues.java
@@ -50,6 +50,6 @@ public class KeyValues {
     }
     
     public static String getStringMod(int i) {
-        return keys[(int) (Math.abs(i % MAXIMUM_ELEMENTS))] ;
+        return keys[(Math.abs(i % MAXIMUM_ELEMENTS))] ;
     }
 }
